Test engine
This patch introduces the test engine, which is one of the core parts of
Rally. The test engine is responsible for processing the tests configuration,
launching verification and benchmarking tests and for processing the results
using Zipkin and Tomograph.

The patch also contains an attempt to carry out some significant refactoring
of the Tester class (which gets initialized and heavily used from the test
engine) and also a couple of additional tests to test the new functionality.

NB: many tests are still to be implemented.