Bugfix for ConfigManagers: defauls overwriting
In the previos version the read_from_dict() method of the test ConfigManager
classes did not overwrite the already existing options by default, which
resulted in the incorrect TestEngine.bind() behavious. This patch fixes this
issue, adds comments to the ConfigManager class and also makes one small
styllistic improvement in TestConfigManager.

Change-Id: Ie5943e6837405719694659d93912b6153601f878
Closes-Bug: #1227864