Rewrite rally.utils.sync_execute() to wait_for()
The patch simplifies the logic of rally.utils.sync_execute(), removing
the function call from this method. It is assumed that the function call
to create the desired resource will be performed by the developers before
calling wait_for(), which would result in a bit of code duplicate but would
significantly improve the overall code readability.

Blueprint benchmark-scenarios

Change-Id: Ieeb365d394b6dad6016047c3b4e6eb0e2543a613