Data processing: aggregation plots
This is the first contribution to the automated data processing and
visualization in Rally. Here we introduce the method for plotting
the relation between the values of some config parameter of a benchmark
scenario and its runing time. Minimal, maximal and average runtimes
are plotted, in assumption that the benchmark scenario has been run
several times during the task execution with different values of the
parameter the data is aggregated on (e.g. the "active_users" parameter).

The patch also extends the CLI with a new command, namely
"task plot aggregated <parameter_name> <task_uuid>"

Blueprint data-processing

Change-Id: Ieee6622f6cfdc737878a0217cf965307d739b2ce