Fully initialize the keystone client in osclients
To be able to perform some administrative tasks like listing/creating tenants
etc., the keystone client needs to use a public address which is not guaranteed
in case of using the "auth_url" parameter only. This patch adds the "endpoint"
parameter to the input of the keystone client constructor + authenticates the
client on this URL so that it gets fully functional.

Change-Id: I50ef23d4f916de8596aa2fea6f1babdb7742798f
Closes-Bug: #1243273