Use random network in boot_server scenario
In this scneario we are going to run a lot of VMs one by one.
Usually one network has about 500 fixed ips (or we will get performance issues)

So if we are going to run 2k vms we should use at least 4 networks.

To simplify configuration of this test. We will by default chose
random network to get ip for vm.

Change-Id: If5d8c52b4f02315be2bcc542baf43cdcc0989316