Fix intermitten py test failure
Fixes an internmitten py test failure related to File
"tests/serverprovider/test_provider.py", line 64, in
test_get_available_providers. The issue causes the gate jobs to fail on
a patch submission and appears to be a timing error with respect to when
the "inner" test classes ProviderA, ProviderB and ProviderC are visable
in the py runtime.

Change-Id: I74555693135cd69f2d5771de24c410c0bed25244