Add sleep_between() method to benchmark scenarios
This patch adds a helper sleep_between() method to the base.Scenario class
thus enabling the subscenarios to use it. The method performs a time.sleep()
call for a random amount of seconds from the given range and also counts the
amount of seconds slept to take them to account while calculating the running
time of a scenario.

Here also the NovaServers.boot_and_delete_server gets changed using the new
sleep_between method(): it can now (optionally) perform a sleep between
booting and deleting a server.

Blueprint benchmark-scenarios

Change-Id: I2a76dadbd8413cef0650c2af7cf5125188cb35ba