Fixes rally boot_server benchmark bug
Fixes rally NovaServers.boot_server benchmark bug which occurs due to
invalid attribute reference of the classes nova client. This bug appears
to have been introduced when rally switched a dict based set of clients.

The updates herein also permit the boot_server benchmark to function if
no networks are created on the Cloud. In this case the boot call does
not pass down a nic to provision.

This commit also includes boot_server unit tests to provide (obviously)
missing coverage for the benchmark method.

Closes-Bug: #1246407

Change-Id: Id8ca25f5060074a57ff97f1bfa54454dcc30654f