Run benchmark scenarios for a given period of time
This patch introduces a new parameter to the benchmark config, namely
the "duration" parameter. In case it is used instead of the "times"
parameter in the config, it specifies the period of time (in minutes)
during which the benchmark should be launched over and over (so that
the actual amount of scenario runs gets undefined).

A benchmark config which would launch booting/deleting servers concurrently
from several users/tenants during 5 minutes would then look like this:

"benchmark": {
    "NovaServers.boot_and_delete_server": [
        {"args": {"flavor_id": 1,
                  "image_id": "73257560-c59b-4275-a1ec-ab140e5b9979"},
         "execution": "continuous",
         "config": {"duration": 5, "active_users": 2,
                    "tenants": 3, "users_per_tenant": 2}}
    ]
}

Blueprint flexible-benchmark-config

Change-Id: Ic5085d973abf177077dcabe57b92992e02459370