Enable access to admin openstack clients from benchmarks
Currently the clients() method of the base benchmark scenario class returns
python openstack clients for one of the temporary created users with non-
administrator permissions. This will be, however, insufficient for developing
benchmark scenarios for keystone, where access to admin permissions is
essential.

The patch adds the admin_clients() method to the base benchmark scenario class.
The method returns the requested python client with admin's permissions which
is initialized in the ScenarioRunner class.

Blueprint benchmark_scenarios

Change-Id: I1792d278bb972c0adcb8745f330f6d06b4bf0688