Remove a task from providers
The patch removes a task from server providers. An initialization method
was added to a base class of server providers. The method takes a two
position arguments: a deployment and a configuration. The class method
get_provider was also changed and has got a similar interface.

All present providers were adapted to the new interface.

blueprint independent-deploy

Change-Id: I1f56b7ce1903d3f51713228df0a4a90ccdb199ca