Cosmetic change to cleanup output
Cleanup failures turned out to be a pretty common thing since there may be
dozens of different cloud configurations and it is not trivial to implement
a single cleanup strategy that would work for all of them. This patch makes
the log output in case of a cleanup failure somewhat less affright.

Change-Id: Ibff9edaada9b8a6d69575d6702ab6ea4f2192f8a
Related-Bug: #1260059