Fix usage of get() methods in python clients
Here, we replace passing the reference to a resource to the get() method of
python client managers by passing the ID of the resource. While the previous
approach worked for Nova, it turned out to be a bug in Cinder.

We also fix the list/dict objects used as the parameter defaults, which is
a bad approach as well.

Change-Id: I9baa805b0aa3f7ccc704a2fa64a057858034977c
Closes-Bug: #1263770