Clarify usage of an endpoint
Some debris were removed in this patch. The debris have been lost when
OSFT verification test was removed. This patch involves a clear endpoint
that represents an installation of OpenStack cluster. The current
format of an endpoint looks like that:

{
    "auth_url": "http://example.com:5000/v2.0",
    "username": "admin",
    "password": "secret",
    "tenant_name": "demo"
}

Each deploy engine should to return an endpoint as a result of
make_deploy method.

TODO: We can instantiate rally.osclients.Clients more simply such as
Clients(**endpoint).

Change-Id: Id5217714d08135b935e4594f931e04c7f04861c5