Base part of a server REST API
The patch introduces a first version of REST API. Rally can be run
as-a-Service by the shell script rally-api. By default a service binds
to a socket 0.0.0.0:8877. The pair of binding parameters have got own
section called an 'api'. One of the supported Content-Type is an
'application/json'.

TODO:
* A Deployment and a Task controllers/types.
* Proper exception handling.

blueprint rally-service-api-v1

Change-Id: Ib0a1f141a083a35c8d43aed2bb1db3307e022b70