Fix mocking in scenario runner tests
The recent patch on scenario runner refactoring involved moving a couple of
mocks back from the decorator syntax inside the test methods (which was a
part of the fix proposed against the hanging python26 unit tests). This seems
to be a mistake which gets fixed in this patch.

Blueprint unit-tests-refactoring

Change-Id: Ib6f084b06e640fd0da4d49ea70c7a5a30a2f0180