Fix CIDR key error bug
The nova.security_groups.create() requires the cidr in the following
format: cidr="0.0.0.0/0", however when the the rule is listed as part of
a security group, it is returned as "ip_range": {"cidr": "0.0.0.0/0"}.

This patch ajdusts fakes to replicate this behaviour so that the tests
can be fixed. Also adjusts the criteria in the rules_to_add dictionary
to be correct.

Change-Id: Ifaa0e92fdc240e4aea114c062632c5fe539e7921
Closes-Bug: #1286096