Proper fix for test_run_scenario_exception_inside_test
We were getting non deterministic bugs in this method. But adding
more time like here  I1007436736bf9d42b98c8782a6141203c271aa75 is
not a proper way.

If we change order of check duration & running method we will be able
to run it exactly one time, even if we set duration == 0.

So it will raise exception => we will be able to check that is raised.

Change-Id: I64964f8386a06f0a54e134d62a59e0955e3a8d75
Closes-bug: #1283372