exclude INFO log about "Starting new HTTP connection..."
The request.urllib3 module always throw INFO level logs onto screen.
That's noise (we are not only people complain this [0]). For the reason
we use logging.setup method as settings, we could change the root logger
is None instead of rally. So this patch just try a 'hack' way to avoid
those logs by set a higher level when user didn't specify a log
configuration.

[0] http://stackoverflow.com/questions/11029717/how-do-i-disable-log-messages-from-the-requests-library

Change-Id: Ib637c04984115bbdf79a237948cca480759f96f4