improve rally output (exception part)
This patch does a lot of trivial changes. But all changes belong to two
things.

The first thing is to make clear that when we could put traceback in rally
output. There is a strategy:

----------------------------------------
|  err type    |  [0]  |  [1]  |  [2]  |
| normal mode  |   N   |   N   |   N   |
| verbose mode |   N   |   N*  |   Y   |
|  debug mode  |   N   |   Y   |   Y   |
----------------------------------------

Why need the [1] between [0] and [2]? The reason is that OpenStack APIs
always return something hard to explain[3]. In another word, we need some
*hacking* for it in rally codes. At least we have to parse it in a way
different with [0] and [2].

The second thing is improving error message so that we could understand
output better. An easy example is that when updating resource status,
the old output always said "resource is in ERROR status". Who is it?
What's the reason for it? This is what this patch is trying to tell you.

Firstly some exception message is improved:

GetResourceFailture: telling users *which resource* failed and *why*
GetResourceNotFound: telling users *which resource* is not found
GetResourceErrorStatus: telling users *which resource* is in a *error
                        status* and *why* (if possible)

And print fault message of nova servers.

[0] Server side fault
[1] Client side fault caused by server side
[2] Client side fault
[3] bug #1280453

Change-Id: I3082a72b6547493ed19def83847fe8f6e9ff0e2f