Add histogram to the report of tasks' results
Add two histogram charts to the set of charts of the HTML report
produced by the 'rally task plot2html' command. The first histogram
depicts the whole task, whereas the second histogram depicts every
atomic action of the task.
For every of these two histograms, there is a drop down menu from which
the user can select the type of the histogram. The type of the histogram
is described by the method that was used to calculate the number of
bins. At the moment, there are four types and it is easy to add more.
In general, a histogram has on its X axis a continuous variable
separated in a number of bins (specific ranges) and on its Y axis the
frequency of events for every bin.
In Rally's case, the histogram has on its X axis the duration of a
benchmark scenario in seconds, separated in a number of bins, and on its
Y axis the number of iterations that fall into every defined bin.

Change-Id: Ibb8d9a7bab40a00adff25df33bac6bb0e71dfe41
Implements: blueprint cli-task-plot2html-histogram