Name Change for "active_users" to "concurrency"
There would be two kind of "users" appears in the task
configuration file when we're using runner "constant".
One in "context" means the number of openstack users,
while the other in "runner" simulate the activities
of multiple "human users", which actually controls the
number of concurrent scenarios.

Rally user might easily be misleaded.

Change-Id: I4333885d877364fbd19279cf338b758abbb804f7