Enable plots with no successful iterations
The stacked and pie graphs can still be produced when there are no
successful iterations. The pie chart ilustrates this well by showing 100%
error. However the histogram cannot be produced as the number of bins
cannot be calculated, so this is skipped.

Also hides atomic actions header and select when there are no atomic
actions results.

closes-bug: #1315063
Change-Id: Ic1fb7f00550828a8e374a94034c1b8eb07db5e89