Add "rally info" command
Here we add a new "rally info" command, which, based on the input,
prints descriptions for different entities in Rally (in this patch -
benchmark scenarios & scenario groups, more entities are to follow
(deployment engines etc.)). This command makes use of docstrings by
parsing them and printing them in CLI in human-readable form.

The command should be called as "rally info find <query>" (or, in case
of an explicit parameter setting, "rally info find --query=<query>").

Samples:

    $ rally info find create_meter_and_get_stats
    CeilometerStats.create_meter_and_get_stats (benchmark scenario).

    Test creating a meter and fetching its statistics.

    Meter is first created and then statistics is fetched for the same
    using GET /v2/meters/(meter_name)/statistics.

    Parameters:
        - name_length: length of generated (random) part of meter name
        - kwargs: contains optional arguments to create a meter

    $ rally info find Authenticate
    Authenticate (benchmark scenario group).

    This class should contain authentication mechanism.

    For different types of clients like Keystone.

Change-Id: Icf3545c0666d99ab7fd0eaabce8bbe572834e485