Fix semantic validation of context images
In the semantic validation of the images context it was assumed that the
parameter users is a dictionary that contains the key endpoint, whereas
users is a list of Clients objects. This was leading to a validation
error and any benchmark scenario that had in its configuration file this
context was failing.

Closes-Bug: 1356336
Change-Id: I6452a47975cd3adcd31ec7b7c4ca0159aec8776f