Add simple multi thread broker
To make creation/deletion of resources in different context it's very usefull
to use broker module. Where we are publishing to queue "tasks" and consuming them
from different threads.

In case of working with OpenStack resources we should cache initialized clients
and initialized clients can't be shared between threads.

So to be able to cache clients each thread has it's own *cache* dict

Co-authored-by: Mikhail Dubov <mdubov@mirantis.com>
Change-Id: Ic67bd45df375de98140c3fdd20a1af129f0886f0