Refactor generic cleanup mechanism part 2
Current cleanup mechanism is awful. It's hardcoded, without repeat on failures,
contains a lot of mistakes, in some places it is too fast (in case of deletion of VMs)
in other to slow (like deletion of users). As well there is mix of cleanup mechanism
and resource cleanup mechanism.
To resolve these all issues this patch introcude a cleanup engine that resolves all issues
above.
It's resource based, so to add new resouce you should just a make subclass of base.Resource
and probably override some of methods like (list, delete, is_deleted) and that's all.
All complexity of managing:
0) waiting until async deletion is finished
1) repeat on failure logic
2) greaceful failure handling
3) parallelization
4) plugin support
Is hidden deep inside cleanup engine
+ bonus we are able to specify now single resource (without clenaping whole service)

PART 2:
-------

* Write cleanup plugins for all tracked resources (on currnet moment)
* Switch cleanap context to use cleanup engine isntead of ald cleanup.utils
* Merge admin and user cleanup context (cause they are quite short now)
* Change quotas scenairos cleanup context configuration
* Fix resource names that are created by ceilometer scenarios

bp benchmark-context-cleanup-refactor

Change-Id: I85ab881fd65da4bb8263b4f72b9f751673ae5099