Changing default new user password
Due to some security reasons, this patch changes default user assignment
from "same as name" to uuid.

Also, named parameter "password" is removed from method signature.
If there is a need for a new user to be created with preselected
that can be done thru kwargs by seting value with the "password" key.

Change-Id: Iccd085a8e67b204d4f9a985d5628128f4299ff83