rally info: smart substitutions in the non-ambiguous case
When the user inputs his query just with a couple of typos, however,
there is typically only one substitution suggested by the system:

$ rally info find NovaServers.boot_and_delete_servers
Failed to find any docs for query: 'NovaServers.boot_and_delete_servers'
Did you mean one of these?
    NovaServers.boot_and_delete_server

The idea is that in this case, when there is only one possible substitution,
we could perform this substitution automatically, whithout requesting the user
to do so. In the example above, Rally could response exactly as if the user
typed the correct query (which is NovaServers.boot_and_delete_server, without
“s” at the end):

$ rally info find NovaServers.boot_and_delete_servers
NovaServers.boot_and_delete_server (benchmark scenario).

Tests booting and then deleting an image.

Change-Id: Ie2f8ff52e21c038ce72a77299de063d7ffddf24b