Disable security group for some neutron driver
Some driver may not support security groups, in which
case the driver may return 500 and causes a failure in Rally.

This commit checks whether the driver supports secgroup
when neutron networking is enabled when neutron driver
reports its extensions correctly, or the user may set
allow_secgroup to false in the context setting to disable secgroup.

Change-Id: Id310af070d09bdf21ecb865be11a55b65f0d3837