Avoid mapping non-boot volume to boot device
When running a scenario that does not involve booting from volume, it is
important to avoid mapping the volume to be attached to the boot device.
vda is often a boot device. This uses vdrally which should never be
picked up as a boot device.

There may be a better way to do this by converting from
block_device_mapping to block_device_mapping_v2, where it is possible to
express a boot_index, but that's a more invasive change.

Change-Id: Ib0a4af3a3b2d8cf3db333cdadf312e078ba47075