Add the context custom_image
Add a new context with the name `custom_image'. This context is
hidden and acts as the base for the contexts of the virtual machine
boot from a custom prepared images. This is used for e.g. preparing
an image with benchmark application preinstalled.

If there is `admin' user then image is created using the first available
user and then makes it public using the `admin' user.
If no `admin' user is available then image is created for each user.

Co-Authored-By: Tzanetos Balitsaris <tzabal@freebsd.org>
Implements: blueprint benchmark-vms
Change-Id: I63a6a0f0f8f1270014ef8322b8e1998b162eb760