Add user-friendly message if the task cannot be deleted
If the final status of the task is unfinished(verifying, running) despite
the rally process is not running any more because the rally process was
crashed (For example, killed the rally process), could not delete the task
and outputed confusable trace error message.
Therefore changed it not to output confusable trace error message and
print user-friendly message("Use '--force' option to delete the task with
vague state.").

Change-Id: Ia5c25022b026c683cdfdfcc6f94928d27e39b9f7
Closes-Bug: #1421142