Fix bash autocompletion of "rally task sla_check"
Character "_" is used as delimiter of commands and
their subcommands stored in bash array as keys,
i.e. in form "task_sla_check". Current code splits
it into an array based on delimiter "_" which would
return array with three values - task, sla and check.
The patch takes first value from the beginning to
first "_" as command and the rest as its subcommand.

Change-Id: I7d8c9b1eeb7c28ef24b0f2f366c6a18930282be5
Closes-bug: 1401180