[Sahara] Add support for anti-affinity
The anti-affinity feature allows to schedule one vm per compute in the
cluster. This feature is ment to provide the best reliability for HDFS
service of Hadoop. Also it can be used to "warmup" sahara images which
means the images are going to be cached on the compute node after the
firs launch. All the next benchmarks sohuld run faster with such image.

Change-Id: Ib39b4cff5d8ec525d2af143d5bc22185ff7339ca