Fix rally plugins relative imports
We are not able to use relative imports for plugins that are
outside of rally package.

To fix this we should add to sys.path all directories with Rally
plugins.

This is quite easy fix. But in any case it includes functional test
that checks that this actually works in real life.

+ Fix of Rally gates to copy whole tree not just *.py files from
  rally-jobs/plugins/

Change-Id: Ie53dc2ba5c32a5e83d84d0ea5703e2e04b10ff33
Closes-bug: #1446923