Cleanup heat stacks 1 time per tenant
Heat stacks are tenant-based, it means that when manager is
returning list of stack for the whole tenant(not for user).
So when rally is cleaning up stacks it needs to delete it
only 1 time per tenant. Otherwise, it is very likely to get
"stack not found" errors in log.

Change-Id: Id8b0f967e2a1b35c8de1abdeee40fc060ce2a0d2
Closes-bug: #1450997