Allow for default() filters for jinja variables
The custom checking for undefined variables was incorrectly considering
a variable that uses a default filter as being undefined. A default
filter allows for providing a default value for an otherwise undefined
variable, and thus it should be allowed to pass through the initial
parse. This newly added regex will find default filter usage and allow
it.

Change-Id: I58991b30221036600f4c0cea7ae5473a5e439573