Import tempest results into Rally database
Add a new subcommand to the `verify` command to import tempest results
into an existing deployment in the Rally database. `verify import`
will read a tempest log in subunit format, and import it into the
current deployment (or the one specified in the command line)

The command should be called as

  $ rally verify import --file [log_file] --deployment [uuid]

This version of the patch will require a normal deployment to be
created beforehand.  Also, the log_file needs to be a tempest log in
subunit format; `--file` parameter is required.

Apart from `--file` and `--deployment` parameters, this command
understand `--set-name` and `--no-use` from the `start`
command. `--set-name` can be used to attach a name or a small
description to the tempest result, and `--no-use` will mark the new
verification object as a no-current. So, for example, if we want to
assign a name to the tempest result we can use:

  $ rally verify import --file [log_file] --set-name [name] \
          --deployment [uuid]

Implements: blueprint verification-import

Change-Id: I65194a2f456106fa5731a13f53d5985971aad43a