OpenStackProvider: Purge disabled code that reads arbitrary files
There is some half-existing code in the OpenStack serverprovider for a
"userdata_file" option that reads userdata from a file path.  Currently
this feature doesn't work because:

A) The config schema rejects the userdata_file property.

B) There's a bug where the file contents are only passed to the first
   server created (should use a string rather than a file handle in the
   loop below).

This change removes the vestigial userdata_file code to prevent anyone
being tempted to "fix" this feature and allowing reading of arbitrary
server-side files in some distant future.

If/when this feature is desired, it needs to be clear that the
userdata_file contents will be read on the client side of some
hypothetical client/server split (ie: in the CLI).

Closes-Bug: #1484309
Change-Id: Id0134f7a1087bf62886b889b8a84f1cc5abeeeac