Keystone get_entities scenario backwards compat
This lets the user supply a service name to the Keystone get_entities
scenario in order to fetch a service other than 'keystone,' which
makes that scenario work for older OpenStack. The user can also supply
None, which will cause the scenario to create a new service, just like
it does for tenant, role, and user.

As a side-effect, this fixes the unit test for get_entities.

Change-Id: I3393c9bdbbab37b6d37c24b0bd7701c0dbbbfe75