Implement subunit_parser
subunit2json module was implemented based on subunit2html module, which
cause several issues(see attached bugs for more details).

The main approach of new parser: it designed to be used from Python code,
not as live-filter for a stream.

Since such parser can be used not only for verification, the implementation
is moved to common dir.

Also, this patch:
 - removes fake log, which was left from xml parser;
 - removes all db calls(except list call) from cli.verify module.

Closes-Bug: #1456810
Closes-Bug: #1471400
Change-Id: Ifcfce827a0df1e75f58e198ee899d4bb85ca9376