[Reports] Add class for merging data stream
This patch adds class rally.task.processing.utils.GraphZiper
that can process unlimited stream of some values (assume
that is a kind of iteration value) and compose a result list
with expected length.

This is useful for processing data for reports charts, so we
can process arbitrary count of iterations without memory load
and generate effective chart for these iterations.

Change-Id: Iec28b26fcbeb8523bb0b01cf0723be37c76f0127