Add ANSI color to heredoc output
Install rally script uses heredoc (cat < __EOF__) in lots of
places so to make output more verbose, added bash ansi colors
to it. The warnings and errors are formatted in RED while options
and success message is shown in GREEN. After formatted output,
NO_COLOR is applied to change it back to normal.
For more info on bash ansi colors, see:
http://misc.flogisoft.com/bash/tip_colors_and_formatting

Change-Id: Ic105c3938137fe29359d2a4e543e47a95829f777