[Reports] Small simplification in new HTML charts classes
This patch simplifies and unifies render() return value format
for some HTML report classes, so format will be changed at
JavaScript part. That is because there will be another data
transformations (like setting of specific bar color) on frontend
and there is no reason to split this processing between Python
and JavaScript.

Change-Id: I7764babaa53db618e90eb15d08353a48f699188d